requirements.txt
cycler==0.10.0
dask==0.11.0
decorator==4.0.10
Django==1.10
image==1.5.4
mahotas==1.4.1
matplotlib==1.5.2
networkx==1.11
numpy==1.11.1
Pillow==3.3.1
pyparsing==2.1.8
python-dateutil==2.5.3
pytz==2016.6.1
scikit-image==0.12.3
scikit-learn==0.17.1
scipy==0.18.0
SimpleCV==1.3
six==1.10.0
toolz==0.8.0
